requirements.txt
mediapy
tqdm

# torch==1.10.0
# torchvision==0.11.1
pytorch-lightning==1.5.2
einops==0.3.2
kornia==0.6.1
opencv-python==4.5.4.58
matplotlib==3.5.0
jupyter
imageio==2.10.4
imageio-ffmpeg==0.4.5
torch_optimizer==0.3.0

# for mesh
PyMCubes==0.1.2
pycollada==0.7.1
trimesh==3.9.1
pyglet==1.5.10
networkx==2.5

# for point cloud
plyfile==0.7.2
open3d==0.13.0
